requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
DateTime==4.4
Django==4.0.2
Pillow==9.0.1
pytz==2021.3
sqlparse==0.4.2
tzdata==2021.5
zope.interface==5.4.0
